+++ packages/bluesky/lib/src/entities/list_view.freezed.dart
@@ -20,6 +20,8 @@ ListView _$ListViewFromJson(Map<String, dynamic> json) {
 
 /// @nodoc
 mixin _$ListView {
+  @JsonKey(name: '\$type')
+  String get type => throw _privateConstructorUsedError;
   String get purpose => throw _privateConstructorUsedError;
   @AtUriConverter()
   AtUri get uri => throw _privateConstructorUsedError;
@@ -45,7 +47,8 @@ abstract class $ListViewCopyWith<$Res> {
       _$ListViewCopyWithImpl<$Res, ListView>;
   @useResult
   $Res call(
-      {String purpose,
+      {@JsonKey(name: '\$type') String type,
+      String purpose,
       @AtUriConverter() AtUri uri,
       String cid,
       String name,
@@ -73,6 +76,7 @@ class _$ListViewCopyWithImpl<$Res, $Val extends ListView>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
+    Object? type = null,
     Object? purpose = null,
     Object? uri = null,
     Object? cid = null,
@@ -85,6 +89,10 @@ class _$ListViewCopyWithImpl<$Res, $Val extends ListView>
     Object? indexedAt = null,
   }) {
     return _then(_value.copyWith(
+      type: null == type
+          ? _value.type
+          : type // ignore: cast_nullable_to_non_nullable
+              as String,
       purpose: null == purpose
           ? _value.purpose
           : purpose // ignore: cast_nullable_to_non_nullable
@@ -153,7 +161,8 @@ abstract class _$$_ListViewCopyWith<$Res> implements $ListViewCopyWith<$Res> {
   @override
   @useResult
   $Res call(
-      {String purpose,
+      {@JsonKey(name: '\$type') String type,
+      String purpose,
       @AtUriConverter() AtUri uri,
       String cid,
       String name,
@@ -181,6 +190,7 @@ class __$$_ListViewCopyWithImpl<$Res>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
+    Object? type = null,
     Object? purpose = null,
     Object? uri = null,
     Object? cid = null,
@@ -193,6 +203,10 @@ class __$$_ListViewCopyWithImpl<$Res>
     Object? indexedAt = null,
   }) {
     return _then(_$_ListView(
+      type: null == type
+          ? _value.type
+          : type // ignore: cast_nullable_to_non_nullable
+              as String,
       purpose: null == purpose
           ? _value.purpose
           : purpose // ignore: cast_nullable_to_non_nullable
@@ -242,7 +256,8 @@ class __$$_ListViewCopyWithImpl<$Res>
 @JsonSerializable(includeIfNull: false)
 class _$_ListView implements _ListView {
   const _$_ListView(
-      {this.purpose = 'app.bsky.graph.defs#modlist',
+      {@JsonKey(name: '\$type') this.type = 'app.bsky.graph.defs#listView',
+      this.purpose = 'app.bsky.graph.defs#modlist',
       @AtUriConverter() required this.uri,
       required this.cid,
       required this.name,
@@ -257,6 +272,9 @@ class _$_ListView implements _ListView {
   factory _$_ListView.fromJson(Map<String, dynamic> json) =>
       _$$_ListViewFromJson(json);
 
+  @override
+  @JsonKey(name: '\$type')
+  final String type;
   @override
   @JsonKey()
   final String purpose;
@@ -292,7 +310,7 @@ class _$_ListView implements _ListView {
 
   @override
   String toString() {
-    return 'ListView(purpose: $purpose, uri: $uri, cid: $cid, name: $name, description: $description, descriptionFacets: $descriptionFacets, avatar: $avatar, createdBy: $createdBy, viewer: $viewer, indexedAt: $indexedAt)';
+    return 'ListView(type: $type, purpose: $purpose, uri: $uri, cid: $cid, name: $name, description: $description, descriptionFacets: $descriptionFacets, avatar: $avatar, createdBy: $createdBy, viewer: $viewer, indexedAt: $indexedAt)';
   }
 
   @override
@@ -300,6 +318,7 @@ class _$_ListView implements _ListView {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$_ListView &&
+            (identical(other.type, type) || other.type == type) &&
             (identical(other.purpose, purpose) || other.purpose == purpose) &&
             (identical(other.uri, uri) || other.uri == uri) &&
             (identical(other.cid, cid) || other.cid == cid) &&
@@ -320,6 +339,7 @@ class _$_ListView implements _ListView {
   @override
   int get hashCode => Object.hash(
       runtimeType,
+      type,
       purpose,
       uri,
       cid,
@@ -347,7 +367,8 @@ class _$_ListView implements _ListView {
 
 abstract class _ListView implements ListView {
   const factory _ListView(
-      {final String purpose,
+      {@JsonKey(name: '\$type') final String type,
+      final String purpose,
       @AtUriConverter() required final AtUri uri,
       required final String cid,
       required final String name,
@@ -360,6 +381,9 @@ abstract class _ListView implements ListView {
 
   factory _ListView.fromJson(Map<String, dynamic> json) = _$_ListView.fromJson;
 
+  @override
+  @JsonKey(name: '\$type')
+  String get type;
   @override
   String get purpose;
   @override