+++ packages/bluesky/lib/src/entities/list_view_basic.dart
@@ -17,6 +17,9 @@ part 'list_view_basic.g.dart';
 class ListViewBasic with _$ListViewBasic {
   @JsonSerializable(includeIfNull: false)
   const factory ListViewBasic({
+    @Default('app.bsky.graph.defs#listViewBasic')
+    @JsonKey(name: '\$type')
+    String type,
     @Default('app.bsky.graph.defs#modlist') String purpose,
     @AtUriConverter() required AtUri uri,
     required String cid,