+++ packages/bluesky/lib/src/entities/list_view_basic.freezed.dart
@@ -20,6 +20,8 @@ ListViewBasic _$ListViewBasicFromJson(Map<String, dynamic> json) {
 
 /// @nodoc
 mixin _$ListViewBasic {
+  @JsonKey(name: '\$type')
+  String get type => throw _privateConstructorUsedError;
   String get purpose => throw _privateConstructorUsedError;
   @AtUriConverter()
   AtUri get uri => throw _privateConstructorUsedError;
@@ -42,7 +44,8 @@ abstract class $ListViewBasicCopyWith<$Res> {
       _$ListViewBasicCopyWithImpl<$Res, ListViewBasic>;
   @useResult
   $Res call(
-      {String purpose,
+      {@JsonKey(name: '\$type') String type,
+      String purpose,
       @AtUriConverter() AtUri uri,
       String cid,
       String name,
@@ -66,6 +69,7 @@ class _$ListViewBasicCopyWithImpl<$Res, $Val extends ListViewBasic>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
+    Object? type = null,
     Object? purpose = null,
     Object? uri = null,
     Object? cid = null,
@@ -75,6 +79,10 @@ class _$ListViewBasicCopyWithImpl<$Res, $Val extends ListViewBasic>
     Object? indexedAt = null,
   }) {
     return _then(_value.copyWith(
+      type: null == type
+          ? _value.type
+          : type // ignore: cast_nullable_to_non_nullable
+              as String,
       purpose: null == purpose
           ? _value.purpose
           : purpose // ignore: cast_nullable_to_non_nullable
@@ -124,7 +132,8 @@ abstract class _$$_ListViewBasicCopyWith<$Res>
   @override
   @useResult
   $Res call(
-      {String purpose,
+      {@JsonKey(name: '\$type') String type,
+      String purpose,
       @AtUriConverter() AtUri uri,
       String cid,
       String name,
@@ -147,6 +156,7 @@ class __$$_ListViewBasicCopyWithImpl<$Res>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
+    Object? type = null,
     Object? purpose = null,
     Object? uri = null,
     Object? cid = null,
@@ -156,6 +166,10 @@ class __$$_ListViewBasicCopyWithImpl<$Res>
     Object? indexedAt = null,
   }) {
     return _then(_$_ListViewBasic(
+      type: null == type
+          ? _value.type
+          : type // ignore: cast_nullable_to_non_nullable
+              as String,
       purpose: null == purpose
           ? _value.purpose
           : purpose // ignore: cast_nullable_to_non_nullable
@@ -193,7 +207,8 @@ class __$$_ListViewBasicCopyWithImpl<$Res>
 @JsonSerializable(includeIfNull: false)
 class _$_ListViewBasic implements _ListViewBasic {
   const _$_ListViewBasic(
-      {this.purpose = 'app.bsky.graph.defs#modlist',
+      {@JsonKey(name: '\$type') this.type = 'app.bsky.graph.defs#listViewBasic',
+      this.purpose = 'app.bsky.graph.defs#modlist',
       @AtUriConverter() required this.uri,
       required this.cid,
       required this.name,
@@ -204,6 +219,9 @@ class _$_ListViewBasic implements _ListViewBasic {
   factory _$_ListViewBasic.fromJson(Map<String, dynamic> json) =>
       _$$_ListViewBasicFromJson(json);
 
+  @override
+  @JsonKey(name: '\$type')
+  final String type;
   @override
   @JsonKey()
   final String purpose;
@@ -223,7 +241,7 @@ class _$_ListViewBasic implements _ListViewBasic {
 
   @override
   String toString() {
-    return 'ListViewBasic(purpose: $purpose, uri: $uri, cid: $cid, name: $name, avatar: $avatar, viewer: $viewer, indexedAt: $indexedAt)';
+    return 'ListViewBasic(type: $type, purpose: $purpose, uri: $uri, cid: $cid, name: $name, avatar: $avatar, viewer: $viewer, indexedAt: $indexedAt)';
   }
 
   @override
@@ -231,6 +249,7 @@ class _$_ListViewBasic implements _ListViewBasic {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$_ListViewBasic &&
+            (identical(other.type, type) || other.type == type) &&
             (identical(other.purpose, purpose) || other.purpose == purpose) &&
             (identical(other.uri, uri) || other.uri == uri) &&
             (identical(other.cid, cid) || other.cid == cid) &&
@@ -244,7 +263,7 @@ class _$_ListViewBasic implements _ListViewBasic {
   @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(
-      runtimeType, purpose, uri, cid, name, avatar, viewer, indexedAt);
+      runtimeType, type, purpose, uri, cid, name, avatar, viewer, indexedAt);
 
   @JsonKey(ignore: true)
   @override
@@ -262,7 +281,8 @@ class _$_ListViewBasic implements _ListViewBasic {
 
 abstract class _ListViewBasic implements ListViewBasic {
   const factory _ListViewBasic(
-      {final String purpose,
+      {@JsonKey(name: '\$type') final String type,
+      final String purpose,
       @AtUriConverter() required final AtUri uri,
       required final String cid,
       required final String name,
@@ -273,6 +293,9 @@ abstract class _ListViewBasic implements ListViewBasic {
   factory _ListViewBasic.fromJson(Map<String, dynamic> json) =
       _$_ListViewBasic.fromJson;
 
+  @override
+  @JsonKey(name: '\$type')
+  String get type;
   @override
   String get purpose;
   @override