+++ packages/bluesky/lib/src/entities/list_view_basic.g.dart
@@ -13,6 +13,8 @@ _$_ListViewBasic _$$_ListViewBasicFromJson(Map json) => $checkedCreate(
       json,
       ($checkedConvert) {
         final val = _$_ListViewBasic(
+          type: $checkedConvert(r'$type',
+              (v) => v as String? ?? 'app.bsky.graph.defs#listViewBasic'),
           purpose: $checkedConvert(
               'purpose', (v) => v as String? ?? 'app.bsky.graph.defs#modlist'),
           uri: $checkedConvert(
@@ -27,10 +29,12 @@ _$_ListViewBasic _$$_ListViewBasicFromJson(Map json) => $checkedCreate(
         );
         return val;
       },
+      fieldKeyMap: const {'type': r'$type'},
     );
 
 Map<String, dynamic> _$$_ListViewBasicToJson(_$_ListViewBasic instance) {
   final val = <String, dynamic>{
+    r'$type': instance.type,
     'purpose': instance.purpose,
     'uri': const AtUriConverter().toJson(instance.uri),
     'cid': instance.cid,